rename top level folder for tests
